+++ README.md
@@ -47,4 +47,4 @@
   - Input Field 를 벗어났을 때 위 조건이 충족되지 않으면 에러 메시지를 띄움
 - [ ] 로그인 버튼
   - ID, PW 조건이 모두 충족되었을 때만 버튼이 활성화
-- [ ] 폼 검증 로직을 커스텀 훅으로 리펙토링
+- [x] 폼 검증 로직을 커스텀 훅으로 리펙토링
